Some fixes
Fix assets icon when we run the cli from another directory then its
source code.

Run notify-send in the background to not block new messages until the
user clicks on the latest message or it was dismissed by the notify
agent.

Only enable notify-send for the linux operating system by default.

Rename userEditor to getUserEditor